chore: remove support for triplet loss variants (e.g. unsupervised)
